requirements.txt
streamlit
openai
autogen-agentchat
autogen-ext[openai]
python-dotenv
google-generativeai
retry
matplotlib>=3.8.0
wordcloud